+++ packages/ui/src/index.ts
@@ -31,7 +31,6 @@ export {
 export const RichTextInput = defineAsyncComponent(() => import('./components/RichTextInput.vue'))
 export type { RichIcon, RichTextLabels } from './components/RichTextInput.vue'
 export { default as ImageUpload } from './components/ImageUpload.vue'
-export { default as NumberInput } from './components/NumberInput.vue'
 export { default as MultiSelect } from './components/MultiSelect.vue'
 export { default as FontUpload } from './components/FontUpload.vue'
 export { default as BaseModal } from './components/BaseModal.vue'